#74: Add unit test to test tabular context filters
